fix: serivce 오타 수정
