+++ backend/src/main/java/com/morjo/model/service/UserService.java
@@ -17,4 +17,20 @@ public class UserService {
         
         return user != null;
     }
+    
+    public boolean join(User user) {
+        // !TODO nickname 유효성 검사
+        // !TODO 다양한 예외 상황들을 어떻게 http 응답에 보낼까
+        return userDao.insertUser(user) == 1;
+    }
+
+    public long getUserByKakaoId(Long kakaoId) {
+        User user = userDao.selectUserByKakaoId(kakaoId);
+
+        if (user == null) {
+            return -1;
+        }
+
+        return user.getUserId();
+    }
 }
